Sesión08/parkwise_project_iter01/identity/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
python-jose[cryptography]

